Organize tool call tests
Generated-by: claude-code Sonnet 4.6
